relative paths and require-jquery until we have nested config
The solution before this commit is what we want, but means that all
dependency config needs to be replicated in applications using the
library. Until we understand how we can reuse the config from
patterns, we switch back to relative paths and require-jquery, so we
do not have to worry about declaring the jquery dependency for jquery
plugins.