Another refactor. See below.
Previously the way setPosition calculated the proper height for a modal was to
clone it and remove the height css attr from the clone. I assume the theory
being that the clone will automatically resize to its full height and from that
the height of the modal can be calculated. In practice this didn't always work.

Now, instead of making a clone, I calculate the height by getting the height of
the tallest element inside the modal. This simplifies things a bit (removes
need for two separate methods).

I also debounce the calls to setPosition.