Enable "source: content" to use an element other than the trigger
Currently, if we set "source: content", the tooltip contents will be the
children of the triggering element.

This patch looks whether the triggering element has an "href" attribute or
"url" parser argument (with a URL containing a fragment identifier) and uses
that to identify the element in the page which should function as the source.