Conform to Jasmine conventions around naming of test methods
Behavior-driven tests, when written with end-user readability in mind, can be
easy-to read and will resemble user-stories or a usage specification.

This means "finishing the sentence" after the "it(" or "describe(" methods.
I.e. taking the names of those methods as the start of a sentence, that is
completed with the spec name.