IE10/11 Bugfix. fixes #379
IE 10/11 was failing because an ajax request was being made with the url
paramater being ''.

Chrome et al. will in this case request the current page, but IE instead
throws a SyntaxError.

However, if the url is '', then we are actually intending to inject content
from the current page, which means that an ajax request is unnecessary.

So the fix is to fetch the content from the current page instead of making an
ajax request.