Refactored testing.
Testing now no longer require a tests bundle to be generated.
Instead tests are run with require.js which loads all files individually.

Some implementation notes:

I reused the requirejs.config of demo/main.js so that the paths don't have to
be specified twice.

However, in order to do this I had to move demo/main.js to ./main.js and
tests/TestRunner.html to tests.html

This means testing in the browser is now at /tests.html and not
/demo/TestRunner.html