Add a new argument "hooks"
This is for other patterns which need to know when injection happened which
concerns them.

The current usecase is pat-raptor.

pat-raptor needs to know when injection happened that saved the form so that
the "Unsaved changes" popup can be suppressed.

The onus is on the designer to put "hooks: raptor;" on the correct pat-inject
element, e.g. the form which will save raptor's contents.