Add a new parser method "addAlias".
The point of addAlias is to add aliases to parser arguments. Aliases will be
parsed under the original name and the options returned by the parser will
contain the original name and not the alias.

This feature is useful when wanting to support both the US and UK variants of
words.