Refactored pat-inject a bit and fixed a bug.
This is the bug: https://github.com/ploneintranet/ploneintranet.prototype/issues/74

The first element being injected is a text node. The code then wrongly triggers
"patterns-injected" on the parent form element and then pat-inject gets
initialized again, which causes event handlers to be registered again.

The solution is to check how many injected elements there are, and not just the
first one. Also, in the registry where injected content is scanned, don't scan
text nodes and scan not on the event.target (which could be the parent), but on
the injected element (which we now provide).

The refactoring was mostly to place long blocks of code in their own methods.