pat-selectbox: adding data-option-value attribute
Currently pat-select (Or perhaps the script is actually in another part of
Patterns) handles that the content of the currently selected option is copied
over to a `data-option` tag on the label of the select or an outjected span.

Normally we use these to create styled selects with a dynamic width in all
browsers with some CSS trickery. We also use it to give colours to the label,
depending on the selected state. Example is workflow states in PI. This is not
a very solid solution as it would required each translated language to be added
to the CSS. This is currently the trouble with PI. Question: could I also get
a data-option-value next to data option? With the value of the value attribute
of the currently selected option in it. As values are typically not translated,
it would solve a problem.