Sexier demo for injection
@pilz @jcbrand , Our sexiest Pattern had the least sexy demo and I
decided to create a nicer one for it. It doesn't cover all the
component can do, but that would be too much for 1 demo anyway and the
documentation is for that. One problem: Equaliser doesn't update after
injection. I believe that led to a problem also somewhere in PI, but
can't remember where. @jc Could you have a look at it? Currently you
need the resize the viewport to be able to read the full poem :).