Add a new argument for inject: confirm
This is used to configure whether a confirmation prompt should be shown to the
user before injection happens.

4 values are available: always, never, class and form-data.

form-data checks if form elements have values. We could also concievably add
something like form-dirty, which checks only if those values have changed, but
this is not yet required currently.

class checks for the 'is-dirty' class on the target.