Add new inject property: loading-class.
Previous the pattern was hardcoded to add the .injecting class to all injection
targets. This can be used to show a spinner, but if you have multiple injection
targets you see multiple spinners, which may be undesirable.

This option can be used to determine whether a loading class should be shown or
not.