Completely redo the scrolling algorithm.
This is an API change and requires the presence of a position:relative
wrapper immediately within the scrollable overflow:auto container.
I've tried to inject that wrapper automatically, but jQuery does not
detect that as a new positioning context and does not properly update
the dom.

The actual calculation is difficult and requires special markup.

Immediately *within* the scrollable, around all the scrollable content,
we need a special wrapper div that provides a positioning context
by being position:relative. Let's call this .inner-scrollable.

This enables us to measure the scrolling distance between the #target
and the top of the scrollable, represented by .inner-scrollable.
We cannot measure directly against the scrollable itself, because if
the scrollable is scrolled down, there's invisible (overflowed) content
'sticking out' *above* the top of the scrollable.

Using the .inner-scrollable top enables us to measure this top overflow.
It doesn't matter whether the scrollable itself provides a positioning
context, since we cannot use that in our calculations anyway.

Because we use .position() which measures the offset versus the positioning
context, we need to walk up the dom across all nested positioning contexts
until we reach the (logical) top of the scrollable: .inner-scrollable.
We MUST ignore the offset of .inner-scrollable versus the scrollable itself!!!