feat(pat inject): Better error handling.
Error handling: In a case of injection error, search for a configurable CSS id selector in the error response.
If it is found, inject the rendered error page into the document.
If no valid error response is found fall back to pre-configured error pages.
This allows for some more informative error pages than a very generic one.
The CSS id selector is defined by an optional URL fragment in the pre-configured error pages.