feat(webpack): Only minimize in production mode.
Compile development and production bundles with the name bundle.min.js.
This allows to get rid to adapt the script name in production and development - both modes use the same name.
However, it's only minified in production mode.
We think the name .min.js also fits the development bundle as it is still babel-transpiled and webpack compiled.