fix(pat toggle): Don't let pat-toggle steal the click event and prevent double clicks.
Do not let pat-toggle prevent the click event after it was catched.
This makes pat-toggle work together with other Patterns, like pat-checklist.
Also, when clicking on a label wrapping a checkbox, the checkbox also emits a click event which results in a unmodified toggle-state.
Prevent double-clicks by debouncing and canceling events.