breaking(core registry): Use a global pattern registry.
Previously it was possible to use multiple pattern registries when multiple
instances of the registry were used, e.g. in different bundles.

With this change we do use only one pattern registry storage which is
shared across instances of ``core.registry`` in multiple bundles.

This is a transparent change - the Patternslib registry API has not
changed at all. Just keep using ``registry.scan``,
``registry.register``, etc.

But if necessary, there is a ``PATTERN_REGISTRY`` export in ``core.registry``
which points to ``window.__patternslib_registry``, a singleton in the global
namespace which is shared across instances of the registry.