fix(Build): Fix changelog template.
Since conventional-changelog 4.2.0 the previous header is merged to the
end of the new changelog entry block.
This is documented here: https://github.com/release-it/conventional-changelog/issues/46
As long as this issue is not resolved, we're adding a ``#`` at the end
of the changelog block, intending the previous entry in the heading
hierarchy, but that's better than a even more messed up changelog file.