breaking(pat navigation): Improve performance by removing the unnecessary mutation observer.
Since we're almost always using pat-inject for replacing or adding DOM nodes
and we already have support for pat-inject here, the mutation observer is not
necessary. Removing it improves the performance in situations where the
navigation structure is updated - for example off-canvas navigation updates
with pat-tabs would invoke many mutation observer callback hits. The previous
performance improvement solved the performance penalty by deferring the
callback for 10ms, but this is taking that further by avoiding it at all.