feat: Add class based base pattern.
This is a new way to define Patterns by extending the BasePattern class and instantiating it on an element.
This is additionally to the old Base pattern approach or even the very old approach of just defining a Pattern following the specs.
The BasePattern class in core/basepattern uses JavaScript classes with all the object oriented benefits.
Yes, that might be syntactic sugar, but then again not.

For usage see: src/core/basepattern.md