feat(pat validation): Validate whole form on submit or single error.
Validate the whole form when a single error happens and some elements were
disabled or when a form submit is done.
This gives the user a better feedback about any data problems in the
form and allows the user to see any other errors at all since the submit
elements could have been disabled and form validation would eventually
not be triggered.