feat(pat-sortable): Add the sortable-item class to each sortable element.
The current situation requires the integrator to add the `sortable-item` class
on all sortables manually. If that is not done there can be styling problems
like a missing insert marker which makes it hard to use the pattern. As the
pattern already defines what should be a sortable it we also let the pattern
assign the class "sortable-item".