fix(pat-date-picker): Do not throw a blur event after selecting a date.
When a date was selected with the date picker a `blur` and `changed`
event was thrown. Now we only throw a `changed` event if da date was
selected. If no date was selected we throw a `blur` event allow
pat-validation do validate required date input fields when no value was
given.