feat(core basepattern): Provide the parser as static attribute.
This change is backwards compatible Change the parser attribute to a
static attribute and provide it also on the object. This change was
necessary because pat-inject was using the parser on a registered
Pattern class to rebase URL configurations in the rebaseHTML method.
There was no access to the parser attribute on non-instatiated objects
before, now it is.