fix(core events): await_pattern_init - check for event coming from correct element.
Check if the init/not-init events were thrown from the Pattern's own
element. When a child element did an unsuccessful Pattern init (rejected
because already initialized) and at the same time the parent element also tried
to initialized the same Pattern await_pattern_init could fail. The not-init
event bubbled up which was incorrectly catched by await_pattern_init on the
parent element.