maint(pat-inject): Remove obsolete hooks option.
The hooks option allowed to throw custom events after successful
injection. It was a multi-value argument but only allowed "raptor" as
value. Raptor was a WYSIWYG editor which has not been further developed
since 10 years and which we're not supporting anymore since quite some
time. Thus this option could be safely removed and this change is not a
breaking change.

If you need to react on events, see the documented event list in
pat-inject's documentation.