feat(pat navigation): Implement scroll-trigger-selector option.
Define the CSS selector which is used to find navigation links with hash
URLs. The default is "a[href^='#'].scroll-marker" which would find all anchor
elements which href starts with a "#" sign and have the class
scroll-marker.
The restriction on scroll-marker allows for other hash-urls in the same
navigation - e.g. a pat-tooltip which references a local content.
If you set it to "none" the scroll marker functionality is not activated.