feat(core registry): Implement sort_early Pattern option.
When scanning a document for patterns, resort patterns and set those with a
sort_early property to the beginning of the initialization chain.

NOTE: Only use when necessary. It's not guaranteed that a pattern with
sort_early is set before another pattern with sort_early. Patterns which
are registered later will have a higher chance to be sorted before
others. Last come, first serve.