fix(pat navigation): Fix current marker anchors with child elements.
The links within a pat-navigation structure might have child elements,
e.g. a span within an anchor. In that case the click target is the span
and not the anchor. Apply a fix where a closest anchor of the click
target is searched, which might be the click target itself.