fix(pat inject): Second attempt to fix the scrolling behavior, where the scrolling target could not be found.
The original problem fixed in commit 911b8b8660197d44291c7d5a9537bbb496df1a38
for 9.9.0-beta.1 addressed a problem where the scroll target was not a
direct child of the injected content but some levels deeper. But this
fix broke the situation where the scroll target is a direct child of the
injected content. The fix applied here handles both situations.