feat(pat-stacks): Add configurable scroll support.
Add "scroll-selector" and "scroll-offset" arguments to pat-stacks,
likewise as in pat-collapsible. "scroll-selector" accepts a CSS selector
string or the special values "self" to scroll to the element itself and
"none" to block any scrolling behavior from parent pat-stacks configuration
options.

With these configuration options you can control the scrolling behavior
for pat-stacks on a fine-grained level.