breaking(pat-scroll): Remove implicit and unpreventable scroll support for pat-stacks.
With a pat-scroll as parent of a pat-stacks enabled navigation anchors,
the page would scroll to the target configured for pat-scroll. This
behavior could not be prevented and led to unwanted effects when using
pat-stacks inside pat-collapsible which itself uses pat-scroll when
configured properly.

This commit removes the scrolling support for pat-stacks within
pat-scroll in favor of an explicit and configurable scrolling behavior
in pat-stacks itself.