feat(core events): remove_event_listener - support wildcard ids.
The id parameter can be a wildcard string, e.g. `test-*-event`, which
would match any event which starts with "test-" and ends with "-event".
The wildcard "*" can be anywhere in the string and also be used multiple
times. If no wildcard is present the search string is used for an exact
match.