fix(pat autosubmit): Trigger the submit event with the correct submitter property.
This resp. the previous commit in event.js fixes a problem where the
submitter was the added to the browser automatically and could lead to
incorrect form submission data in combination with pat-ajax. E.g. we had
a case where for a file upload the delete button's value was submitted,
where it shouldn't have.