feat(core polyfills): Add polyfill for navigation API.
This polyfill adds support for the "navigate" event on the navigation
object. We path "history.pushState" and "history.replaceState" to send
the "navigate" event on the "window.navigation" object when the URL
changes.

This polyfill is for current Firefox and Safari. Chrome based browsers
already support this.

More information on:
https://developer.mozilla.org/en-US/docs/Web/API/Navigation/navigate_event