breaking(pat-depends): change action default to "both"
Set the action default to "both" so that a dependent markup structure is
not only shown respective hidden but contained input elements also enabled and
disabled.

Input fields which are not shown should be disabled, so that they do not
participate in form validation and are not submitted.

If you need a different behavior set it via the "action" parameter on a
pat-depends enabled element itself.