fix(core feature-detection): Fix loading of modernizr script.
In some situations loading of the modernizr.min.js script failed because
the base url could be incorrect. That is fixed with a non-IE compatible
method of getting the current's script URL, which is safe to use.
The problem surfaced in Plone while loading the Patternslib library and
in between the protect.js script was loaded. The base url was calculated
for the protect.js script and the modernizr script could not be loaded.