maint: Remove the public path helper script.
The public path helper script did set the webpack public path to find
the path where the bundle was located and to correctly find the chunk
directory. The logic was based on finding the URL of the last loaded
script. This was not even always correct.

It was also possible to override it via `window.__patternslib_public_path__`.

Since quite some thime, this is not necessary anymore as webpack handles
these path issues on it's own.

This dead code is now removed.