maint: Remove core/remove and core/url.
Remove core/remove and core/url and it's usage in pat-slides.
core/url can nowadays be replaced by the URL interface from the web API.
core/remove did invoke the destroy event to invoke clean-up routines.

core/remove result in an infinite recursion error in combination with the
TinyMCE API due to core/remove's patching of the internal cleanData jQuery API.

pat-slides is currently not in a working state, so the removal is OK.
Anyways, this triggers the question how to handle Pattern cleanup in the
future. The answer is in a mutation observer on document level, but this
will be part of a future release.